renamed default_giskard.py to giskard.py
