+++ src/giskardpy/configs/boxy.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from giskardpy.configs.data_types import ControlModes, TfPublishingModes
-from giskardpy.configs.default_giskard import Giskard
+from giskardpy.configs.giskard import Giskard
 from giskardpy.my_types import Derivatives
