+++ src/giskardpy/configs/hsr.py
@@ -3,7 +3,7 @@ from typing import Optional
 import numpy as np
 
 from giskardpy.configs.data_types import ControlModes, SupportedQPSolver, TfPublishingModes
-from giskardpy.configs.default_giskard import Giskard
+from giskardpy.configs.giskard import Giskard
 from giskardpy.my_types import PrefixName, Derivatives
