+++ src/giskardpy/configs/tiago.py
@@ -3,7 +3,7 @@ from typing import Optional
 import numpy as np
 
 from giskardpy.configs.data_types import SupportedQPSolver, TfPublishingModes
-from giskardpy.configs.default_giskard import Giskard, ControlModes
+from giskardpy.configs.giskard import Giskard, ControlModes
 from giskardpy.my_types import Derivatives
