+++ src/giskardpy/configs/world_config.py
@@ -9,7 +9,7 @@ from std_msgs.msg import ColorRGBA
 
 from giskardpy import identifier
 from giskardpy.god_map import GodMap
-from giskardpy.god_map_user import GodMapUser
+from giskardpy.god_map_user import GodMapWorshipper
 from giskardpy.model.joints import FixedJoint, OmniDrive, DiffDrive, Joint6DOF, OneDofJoint
 from giskardpy.model.links import Link
 from giskardpy.model.utils import robot_name_from_urdf_string
@@ -17,7 +17,7 @@ from giskardpy.model.world import WorldTree
 from giskardpy.my_types import my_string, PrefixName, Derivatives, derivative_map
 
 
-class WorldConfig(GodMapUser):
+class WorldConfig(GodMapWorshipper):
     god_map = GodMap()
 
     def __init__(self):
