+++ src/giskardpy/goals/joint_goals.py
@@ -5,7 +5,7 @@ from typing import Dict, Optional, List
 from geometry_msgs.msg import PoseStamped
 
 from giskardpy import casadi_wrapper as w, identifier
-from giskardpy.configs.default_giskard import ControlModes
+from giskardpy.configs.giskard import ControlModes
 from giskardpy.exceptions import ConstraintException, ConstraintInitalizationException
 from giskardpy.goals.goal import Goal, WEIGHT_BELOW_CA, NonMotionGoal, WEIGHT_ABOVE_CA
 from giskardpy.model.joints import OmniDrive, DiffDrive, OmniDrivePR22
